+++ dice_roller.py
@@ -1,14 +1,17 @@
-from kivy.uix.button import Button
-from kivy.uix.gridlayout import GridLayout
-from kivy.app import App
-from discordwebhook import Discord
-
 import random
 
-import discord_webhook
+from discordwebhook import Discord
+from kivy.app import App
+from kivy.uix.button import Button
+from kivy.uix.gridlayout import GridLayout
 
 
 class DiceButton:
+    """Describes a dimension of dice (for e.g. a Dimension 20 dice [D20])
+    whose number can be increased or decreased,
+    and that can be rolled through calls to its inner, methods which are
+    triggered by with the Object's respective buttons members.
+    """
 
     def __init__(self, dice_size):
         self.dice_to_roll = 1
@@ -23,18 +26,22 @@ class DiceButton:
         self.roll_button.bind(on_release=lambda x: self.do_roll())
 
     def update_label(self):
+        """Changes the label on the main Dice button."""
         self.roll_button.text = f"{self.dice_to_roll}d{self.dice_size}"
 
     def do_increase(self):
+        """Increase the number of dice to be rolled."""
         self.dice_to_roll += 1
         self.update_label()
 
     def do_decrease(self):
+        """Decrease the number of dice to be rolled."""
         if self.dice_to_roll > 1:
             self.dice_to_roll -= 1
             self.update_label()
 
     def do_roll(self):
+
         roll_result_sum = 0
         dice_rolled_results = []
 
@@ -47,8 +54,14 @@ class DiceButton:
         discord.post(content=f"Result : {roll_result_sum} = {self.dice_to_roll}d{self.dice_size} {dice_rolled_results}")
 
 
-class MyApp(App):
+class DiceRoller(App):
     def build(self):
+        """
+        Initialises the GUI for the App.
+
+        :return: (GridLayout) GUI of the Dice Roller App.
+        """
+
         dice_buttons = [
             DiceButton(2),
             DiceButton(4),
@@ -60,7 +73,10 @@ class MyApp(App):
             DiceButton(100),
         ]
 
-        grid = GridLayout(cols=3)  # [-] [dX] [+]
+        """
+            [0: Button for Subtracting Dice | 1: Button for Rolling Dice | 2: Button for Adding Dice]
+        """
+        grid = GridLayout(cols=3)  # [-][dX][+]
 
         for dice in dice_buttons:
             grid.add_widget(dice.decrease_button)
@@ -70,17 +86,28 @@ class MyApp(App):
         return grid
 
 
+def get_webhook(web_hook_file):
+    """Open the file containing the webhook and read the URL,
+    which should be on a single line."""
+
+    try:
+        with open(web_hook_file) as f:
+            return f.read()
+    except FileNotFoundError:
+        print(f"WARNING: The specified webhook file '{web_hook_file}' was not found.")
+
+
 if __name__ == "__main__":
 
-    web_hook_url = discord_webhook.get_webhook("config/web_hook_url.txt")
+    web_hook_url = get_webhook("config/web_hook_url.txt")
     if not web_hook_url:
         print("""
             A web hook URL is needed to start this program.
             See README.md
             """
-            )
+              )
         exit()
 
-    discord = Discord(url=web_hook_url) # Create the Webhook
+    discord = Discord(url=web_hook_url)  # Create the Webhook
 
-    MyApp().run()
+    DiceRoller().run()  # Start the GUI
